Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_cached-expected.txt

Description

This test validates that a 304 Not Modified resource appears in the buffer.


PASS window.performance is defined
PASS window.performance.getEntriesByType is defined
PASS There should be two entries
PASS Both entries should have the same name
